+++ contrib/platform/src/com/sun/jna/platform/win32/WinCrypt.java
@@ -23,8 +23,10 @@
  */
 package com.sun.jna.platform.win32;
 
+import com.sun.jna.FromNativeContext;
 import com.sun.jna.Memory;
 import com.sun.jna.Native;
+import com.sun.jna.NativeMapped;
 import com.sun.jna.Pointer;
 import com.sun.jna.StringArray;
 import com.sun.jna.Structure;
@@ -1194,6 +1196,49 @@ public interface WinCrypt {
         }
     }
 
+    /**
+     * Value wrapper for {@code lpszStoreProvider} of {@link Crypt32#CertOpenStore(CertStoreProviderName, int, HCRYPTPROV_LEGACY, int, Pointer)}
+     */
+    class CertStoreProviderName implements NativeMapped {
+        private final Pointer pointer;
+
+        private CertStoreProviderName(Pointer fixedValue) {
+            this.pointer = fixedValue;
+        }
+
+        public CertStoreProviderName() {
+            pointer = Pointer.NULL;
+        }
+
+        public CertStoreProviderName(int fixedValue) {
+            this.pointer = Pointer.createConstant(fixedValue);
+        }
+
+        public CertStoreProviderName(String name) {
+            byte[] nameBytes = Native.toByteArray(name);
+            this.pointer = new Memory(nameBytes.length);
+            this.pointer.write(0, nameBytes, 0, nameBytes.length);
+        }
+
+        @Override
+        public Object fromNative(Object nativeValue, FromNativeContext fnc) {
+            if (nativeValue == null) {
+                return null;
+            }
+            return new CertStoreProviderName((Pointer) nativeValue);
+        }
+
+        @Override
+        public Object toNative() {
+            return pointer;
+        }
+
+        @Override
+        public Class<?> nativeType() {
+            return Pointer.class;
+        }
+    }
+
     //
     // CryptProtect PromptStruct dwPromtFlags
     //
@@ -1616,6 +1661,68 @@ public interface WinCrypt {
      * certificate when it was exported.
      */
     int PKCS12_INCLUDE_EXTENDED_PROPERTIES = 0x0010;
+
+    //
+    // Store providers for CertOpenStore
+    //
+
+    int CERT_STORE_PROV_MSG = 1;
+    int CERT_STORE_PROV_MEMORY  = 2;
+    int CERT_STORE_PROV_FILE = 3;
+    int CERT_STORE_PROV_REG  = 4;
+    int CERT_STORE_PROV_PKCS7 = 5;
+    int CERT_STORE_PROV_SERIALIZED = 6;
+    int CERT_STORE_PROV_FILENAME_A = 7; // ASCII
+    int CERT_STORE_PROV_FILENAME_W = 8; // Unicode
+    int CERT_STORE_PROV_FILENAME = CERT_STORE_PROV_FILENAME_W;
+    int CERT_STORE_PROV_SYSTEM_A  = 9; // pvPara is ASCII (1 byte/char)
+    int CERT_STORE_PROV_SYSTEM_W  = 10; // pvPara is Unicode (2 bytes/char)
+    int CERT_STORE_PROV_SYSTEM  = CERT_STORE_PROV_SYSTEM_W;
+    int CERT_STORE_PROV_COLLECTION  = 11;
+    int CERT_STORE_PROV_SYSTEM_REGISTRY_A = 12;
+    int CERT_STORE_PROV_SYSTEM_REGISTRY_W = 13;
+    int CERT_STORE_PROV_SYSTEM_REGISTRY = CERT_STORE_PROV_SYSTEM_REGISTRY_W;
+    int CERT_STORE_PROV_PHYSICAL_W = 14;
+    int CERT_STORE_PROV_PHYSICAL = CERT_STORE_PROV_PHYSICAL_W;
+    int CERT_STORE_PROV_SMART_CARD_W  = 15;
+    int CERT_STORE_PROV_SMART_CARD = CERT_STORE_PROV_SMART_CARD_W;
+    int CERT_STORE_PROV_LDAP_W  = 16;
+    int CERT_STORE_PROV_LDAP = CERT_STORE_PROV_LDAP_W;
+
+    //
+    // Store characteristics for CertOpenStore
+    //
+
+    int CERT_STORE_NO_CRYPT_RELEASE_FLAG = 0x00000001;
+    int CERT_STORE_SET_LOCALIZED_NAME_FLAG = 0x00000002;
+    int CERT_STORE_DEFER_CLOSE_UNTIL_LAST_FREE_FLAG = 0x00000004;
+    int CERT_STORE_DELETE_FLAG = 0x00000010;
+    int CERT_STORE_UNSAFE_PHYSICAL_FLAG = 0x00000020;
+    int CERT_STORE_SHARE_STORE_FLAG = 0x00000040;
+    int CERT_STORE_SHARE_CONTEXT_FLAG = 0x00000080;
+    int CERT_STORE_MANIFOLD_FLAG = 0x00000100;
+    int CERT_STORE_ENUM_ARCHIVED_FLAG = 0x00000200;
+    int CERT_STORE_UPDATE_KEYID_FLAG = 0x00000400;
+    int CERT_STORE_BACKUP_RESTORE_FLAG = 0x00000800;
+    int CERT_STORE_MAXIMUM_ALLOWED_FLAG = 0x00001000;
+    int CERT_STORE_CREATE_NEW_FLAG = 0x00002000;
+    int CERT_STORE_OPEN_EXISTING_FLAG = 0x00004000;
+    int CERT_STORE_READONLY_FLAG = 0x00008000;
+
+    //
+    // Store locations for CertOpenStore
+    //
+
+    int CERT_SYSTEM_STORE_CURRENT_USER = 0x00010000;
+    int CERT_SYSTEM_STORE_LOCAL_MACHINE = 0x00020000;
+    int CERT_SYSTEM_STORE_CURRENT_SERVICE = 0x00040000;
+    int CERT_SYSTEM_STORE_SERVICES = 0x00050000;
+    int CERT_SYSTEM_STORE_USERS = 0x00060000;
+    int CERT_SYSTEM_STORE_CURRENT_USER_GROUP_POLICY = 0x00070000;
+    int CERT_SYSTEM_STORE_LOCAL_MACHINE_GROUP_POLICY = 0x00080000;
+    int CERT_SYSTEM_STORE_LOCAL_MACHINE_ENTERPRISE = 0x00090000;
+    int CERT_SYSTEM_STORE_UNPROTECTED_FLAG = 0x40000000;
+    int CERT_SYSTEM_STORE_RELOCATE_FLAG = 0x80000000;
 
     /**
      * Checks for nonfreed certificate, CRL, and CTL contexts. A returned error